// frontend/src/api/generated/models/goalDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * TrainCoach API
 * OpenAPI spec version: v1
 */
import type { GoalPriority } from './goalPriority';

export interface GoalDto {
  id?: string;
  athleteUserId?: string;
  /** @nullable */
  seasonId?: string | null;
  /** @nullable */
  title?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  targetDate?: string | null;
  priority?: GoalPriority;
  isAchieved?: boolean;
  /** @nullable */
  achievedNotes?: string | null;
}